Authoriziation for Front-end

+++ app/Http/Controllers/RoleController.php
@@ -6,6 +6,7 @@ use App\Models\Feature;
 use App\Models\Permission;
 use App\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -14,9 +15,15 @@ class RoleController extends Controller
      */
     public function index()
     {
-        return view('roles.index', [
-            'roles' => Role::all()
-        ]);
+        $permissionId = 6;
+
+        if (Gate::allows('has-permission', $permissionId)) {
+            return view('roles.index', [
+                'roles' => Role::all()
+            ]);
+        } else {
+            return redirect()->back()->with('error', 'Your are not allowed to do this');
+        }
     }
 
     /**
@@ -24,10 +31,16 @@ class RoleController extends Controller
      */
     public function create()
     {
-        return view('roles.create', [
-            'permissions' => Permission::all(),
-            'features' => Feature::all()
-        ]);
+        $permissionId = 5;
+
+        if (Gate::allows('has-permission', $permissionId)) {
+            return view('roles.create', [
+                'permissions' => Permission::all(),
+                'features' => Feature::all()
+            ]);
+        } else {
+            return redirect()->back()->with('error', 'Your are not allowed to do this');
+        }
     }
 
     /**
@@ -35,15 +48,21 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
-        $role = Role::create([
-            'name' => $request->input('role_name'),
-        ]);
+        $permissionId = 5;
 
-        $selectedPermissions = $request->input('permissions');
-        $permissions = Permission::whereIn('id', $selectedPermissions)->get();
-        $role->permissions()->attach($permissions);
+        if (Gate::allows('has-permission', $permissionId)) {
+            $role = Role::create([
+                'name' => $request->input('role_name'),
+            ]);
 
-        return redirect()->route('roles.index')->with('success', 'New Role added successfully.');
+            $selectedPermissions = $request->input('permissions');
+            $permissions = Permission::whereIn('id', $selectedPermissions)->get();
+            $role->permissions()->attach($permissions);
+
+            return redirect()->route('roles.index')->with('success', 'New Role added successfully.');
+        } else {
+            return redirect()->back()->with('error', 'Your are not allowed to do this');
+        }
     }
 
     /**
@@ -59,11 +78,17 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
-        return view('roles.edit', [
-            'role' => $role,
-            'permissions' => Permission::all(),
-            'features' => Feature::all()
-        ]);
+        $permissionId = 7;
+
+        if (Gate::allows('has-permission', $permissionId)) {
+            return view('roles.edit', [
+                'role' => $role,
+                'permissions' => Permission::all(),
+                'features' => Feature::all()
+            ]);
+        } else {
+            return redirect()->back()->with('error', 'Your are not allowed to do this');
+        }
     }
 
     /**
@@ -72,20 +97,26 @@ class RoleController extends Controller
     public function update(Request $request, Role $role)
     {
 
-        $validatedData = $request->validate([
-            'name' => 'required|string',
-        ]);
+        $permissionId = 7;
 
-        try {
+        if (Gate::allows('has-permission', $permissionId)) {
+            $validatedData = $request->validate([
+                'name' => 'required|string',
+            ]);
 
-            $role->update($validatedData);
-            $selectedPermissions = $request->input('permissions');
-            $permissions = Permission::whereIn('id', $selectedPermissions)->get();
-            $role->permissions()->sync($permissions);
+            try {
 
-            return redirect()->route('roles.index')->with('success', 'Role updated successfully!');
-        } catch (\Exception $e) {
-            return redirect()->route('roles.edit', $role->id)->withInput()->with('error', 'An error occurred while updating the role. Please try again.');
+                $role->update($validatedData);
+                $selectedPermissions = $request->input('permissions');
+                $permissions = Permission::whereIn('id', $selectedPermissions)->get();
+                $role->permissions()->sync($permissions);
+
+                return redirect()->route('roles.index')->with('success', 'Role updated successfully!');
+            } catch (\Exception $e) {
+                return redirect()->route('roles.edit', $role->id)->withInput()->with('error', 'An error occurred while updating the role. Please try again.');
+            }
+        } else {
+            return redirect()->back()->with('error', 'Your are not allowed to do this');
         }
     }
 
@@ -95,8 +126,14 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
-        $role->delete();
+        $permissionId = 7;
 
-        return redirect()->route('roles.index')->with('success', 'Role deleted successfully.');
+        if (Gate::allows('has-permission', $permissionId)) {
+            $role->delete();
+
+            return redirect()->route('roles.index')->with('success', 'Role deleted successfully.');
+        } else {
+            return redirect()->back()->with('error', 'Your are not allowed to do this');
+        }
     }
 }
